Refactor: split into files
